vista de carrito
agrega y borra. falta agregar mas cantidades

+++ frontend/zona44app/lib/models/carrito.dart
@@ -9,7 +9,7 @@ class CarritoItem {
     this.cantidad = 1,
   });
 
-  double get precioTotal => producto.precio * cantidad;
+  int get precioTotal => producto.precio * cantidad;
 
   CarritoItem copyWith({
     Producto? producto,
